requirements.txt
Flask==2.3.3
google-api-python-client==2.128.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
openai==1.30.1
python-dotenv==1.0.1
Pillow==10.3.0
playwright==1.43.0
